+++ src/main/java/org/ssutt/core/sql/TTSQLManager.java
@@ -65,4 +65,6 @@ public interface TTSQLManager {
 
     int getLastID(String table) throws SQLException;
 
+    void setQueries(Queries qrs);
+
 }